Decouple protobuf from rest of the core
