+++ sf_core/src/c_api.rs
@@ -1,86 +1,3 @@
-use crate::logging;
-use crate::protobuf_apis::call_proto;
-use proto_utils::ProtoError;
-use std::os::raw::c_char;
-
-#[unsafe(no_mangle)]
-pub extern "C" fn sf_core_init_logger(callback: logging::CLogCallback) -> u32 {
-    let config = logging::LoggingConfig::new(None, false, false);
-    let layer = logging::CallbackLayer::new(callback);
-    match logging::init_logging(config, Some(layer)) {
-        Ok(_) => 0,
-        Err(e) => {
-            eprintln!("Failed to initialize logging: {e:?}");
-            1
-        }
-    }
-}
-
-fn write_buffer(vec: Vec<u8>, buffer: *mut *const u8, len: *mut usize) {
-    let boxed = vec.into_boxed_slice();
-    unsafe {
-        *len = boxed.len();
-        *buffer = Box::into_raw(boxed) as *const u8;
-    }
-}
-
-/// Frees a buffer previously returned by `sf_core_api_call_proto` via `write_buffer`.
-///
-/// # Safety
-/// The caller must pass the exact `buffer` pointer and `len` that were written by a prior
-/// call to `sf_core_api_call_proto`. Each (buffer, len) pair must be freed at most once.
-/// Passing any other pointer or length is undefined behavior.
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn sf_core_free_buffer(buffer: *const u8, len: usize) {
-    if !buffer.is_null() {
-        unsafe {
-            drop(Box::from_raw(std::ptr::slice_from_raw_parts_mut(
-                buffer as *mut u8,
-                len,
-            )));
-        }
-    }
-}
-
-/// # Safety
-/// This function dereferences raw pointers `api`, `method`, `request`, `response`, and `response_len`.
-/// The caller must ensure that `api`, `method`, `request`, `response`, and `response_len` are valid.
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn sf_core_api_call_proto(
-    api: *const c_char,
-    method: *const c_char,
-    request: *mut u8,
-    request_len: usize,
-    response: *mut *const u8,
-    response_len: *mut usize,
-) -> usize {
-    // Prevent unwinding across the FFI boundary. Any panic will be converted to a transport error.
-    let result = std::panic::catch_unwind(|| unsafe {
-        let api = std::ffi::CStr::from_ptr(api).to_string_lossy().to_string();
-        let method = std::ffi::CStr::from_ptr(method)
-            .to_string_lossy()
-            .to_string();
-        let message = std::slice::from_raw_parts(request, request_len);
-        call_proto(&api, &method, message)
-    });
-
-    match result {
-        Ok(Ok(response_vec)) => {
-            write_buffer(response_vec, response, response_len);
-            0
-        }
-        Ok(Err(ProtoError::Application(error_vec))) => {
-            write_buffer(error_vec, response, response_len);
-            1
-        }
-        Ok(Err(ProtoError::Transport(e))) => {
-            write_buffer(e.as_bytes().to_vec(), response, response_len);
-            2
-        }
-        Err(_) => {
-            let msg = b"sf_core panic in sf_core_api_call_proto".to_vec();
-            write_buffer(msg, response, response_len);
-            2
-        }
-    }
-}
+pub use crate::logging::c_api::*;
+#[cfg(feature = "protobuf")]
+pub use crate::protobuf::c_api::*;
