+++ sf_core/src/lib.rs
@@ -16,9 +16,10 @@ mod file_manager;
 pub mod handle_manager;
 pub mod http;
 pub mod logging;
-pub mod protobuf_apis;
-pub mod protobuf_gen;
 pub mod query_types;
 pub mod rest;
 pub mod tls;
 pub mod token_cache;
+
+#[cfg(feature = "protobuf")]
+pub mod protobuf;
