+++ sf_core/tests/e2e/mod.rs
@@ -1,8 +1,13 @@
 #[path = "../common/mod.rs"]
 pub mod common;
 
+#[cfg(feature = "protobuf")]
 pub mod authentication;
+#[cfg(feature = "protobuf")]
 pub mod put_get;
+#[cfg(feature = "protobuf")]
 pub mod query;
+#[cfg(feature = "protobuf")]
 pub mod session;
+#[cfg(feature = "protobuf")]
 pub mod tls;
